+++ src/AggregateFunctions/CMakeLists.txt
@@ -1,21 +1,21 @@
 include("${ClickHouse_SOURCE_DIR}/cmake/dbms_glob_sources.cmake")
 add_headers_and_sources(clickhouse_aggregate_functions .)
-add_headers_and_sources(clickhouse_aggregate_functions combinators)
+add_headers_and_sources(clickhouse_aggregate_functions Combinators)
 
 extract_into_parent_list(clickhouse_aggregate_functions_sources dbms_sources
     IAggregateFunction.cpp
     AggregateFunctionFactory.cpp
-    combinators/AggregateFunctionCombinatorFactory.cpp
-    combinators/AggregateFunctionState.cpp
+    Combinators/AggregateFunctionCombinatorFactory.cpp
+    Combinators/AggregateFunctionState.cpp
     AggregateFunctionCount.cpp
     parseAggregateFunctionParameters.cpp
 )
 extract_into_parent_list(clickhouse_aggregate_functions_headers dbms_headers
     IAggregateFunction.h
-    combinators/IAggregateFunctionCombinator.h
+    Combinators/IAggregateFunctionCombinator.h
     AggregateFunctionFactory.h
-    combinators/AggregateFunctionCombinatorFactory.h
-    combinators/AggregateFunctionState.h
+    Combinators/AggregateFunctionCombinatorFactory.h
+    Combinators/AggregateFunctionState.h
     AggregateFunctionCount.cpp
     FactoryHelpers.h
     parseAggregateFunctionParameters.h
