+++ src/AggregateFunctions/Combinators/AggregateFunctionSimpleState.cpp
