+++ src/AggregateFunctions/Combinators/AggregateFunctionSimpleState.h
