+++ tests/queries/0_stateless/02885_arg_min_max_combinator.reference
